Suez
The Ottomans are threatening the Suez Canal, the bloodline of our empire. First, we must overrun the enemy lines on the East Bank. And from there, swiftly push through the village of Kantara. This will allow us to mount an assault on their artillery positions on Hill 50. When the guns have been silenced, we will push the Ottomans back into the Sinai Desert.